backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Database and Storage
supabase==2.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
cryptography==41.0.8
passlib[bcrypt]==1.7.4

# Image Processing and QR
pillow==10.1.0
qrcode[pil]==7.4.2
opencv-python==4.8.1.78

# AI/ML (with compatible versions)
torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
transformers==4.35.2
sentence-transformers==2.2.2

# Image Forensics
scikit-image==0.22.0
numpy==1.24.4
scipy==1.11.4

# Utilities
requests==2.31.0
aiofiles==23.2.1
pydantic==2.5.0
